requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
cs50==9.4.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Markdown==0.3
Flask-Session==0.8.0
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.12.2
Werkzeug==3.1.3
wheel==0.45.1
